server/requirements.txt
Flask
openai
langchain
langchain-community
langchain-google-firestore
python-dotenv
langchain-openai
bs4
duckduckgo-search
flask-cors
firestore
firebase-admin
requests
langchain-fmp-data
